viewer new design

+++ db/schema.rb
@@ -57,8 +57,8 @@ ActiveRecord::Schema.define(:version => 20130910111438) do
   end
 
   create_table "users", :force => true do |t|
-    t.string   "email",                  :default => "",  :null => false
-    t.string   "encrypted_password",     :default => "",  :null => false
+    t.string   "email",                  :default => "", :null => false
+    t.string   "encrypted_password",     :default => "", :null => false
     t.string   "reset_password_token"
     t.datetime "reset_password_sent_at"
     t.datetime "remember_created_at"
@@ -67,14 +67,14 @@ ActiveRecord::Schema.define(:version => 20130910111438) do
     t.datetime "last_sign_in_at"
     t.string   "current_sign_in_ip"
     t.string   "last_sign_in_ip"
-    t.datetime "created_at",                              :null => false
-    t.datetime "updated_at",                              :null => false
+    t.datetime "created_at",                             :null => false
+    t.datetime "updated_at",                             :null => false
     t.string   "name"
     t.string   "image_file_name"
     t.string   "image_content_type"
     t.integer  "image_file_size"
     t.datetime "image_updated_at"
-    t.decimal  "videoid",                :default => 0.0
+    t.integer  "videoid",                :default => 0
     t.string   "phone"
     t.string   "string"
     t.string   "linkedin"
